Refactor piece initialize into Piece parent class
