+++ week_1/exercises/chess_validator/lib/bishop.rb
@@ -1,11 +1,4 @@
-class Bishop
-	attr_reader :color
-	def initialize(pos_x, pos_y, color)
-		@pos_x = pos_x
-		@pos_y = pos_y
-		@color = color
-	end
-
+class Bishop < Piece
 	def move?(dst_x, dst_y)
 		dx = (dst_x - @pos_x).abs
 		dy = (dst_y - @pos_y).abs
