+++ week_1/exercises/chess_validator/lib/piece.rb
@@ -1,0 +1,9 @@
+class Piece
+  attr_reader :color
+
+  def initialize(pos_x, pos_y, color)
+    @pos_x = pos_x
+    @pos_y = pos_y
+    @color = color
+  end
+end
