+++ week_1/exercises/chess_validator/lib/queen.rb
@@ -1,11 +1,4 @@
-class Queen
-	attr_reader :color
-	def initialize(pos_x, pos_y, color)
-		@pos_x = pos_x
-		@pos_y = pos_y
-		@color = color
-	end
-
+class Queen < Piece
 	def move?(dst_x, dst_y)
 		dx = (dst_x - @pos_x).abs
 		dy = (dst_y - @pos_y).abs
