+++ week_1/exercises/chess_validator/lib/whitepawn.rb
@@ -1,11 +1,4 @@
-class WhitePawn
-	attr_reader :color
-	def initialize(pos_x, pos_y, color)
-		@pos_x = pos_x
-		@pos_y = pos_y
-		@color = color
-	end
-
+class WhitePawn < Piece
 	def move?(dst_x, dst_y)
 		dx = (dst_x - @pos_x)
 		dy = (dst_y - @pos_y)
